Move  to _freesurfer.py
