+++ mne/viz/_3d.py
@@ -21,7 +21,7 @@ import numpy as np
 
 from ..defaults import DEFAULTS
 from ..fixes import _crop_colorbar, _get_img_fdata, _get_args
-from .._freesurfer import _read_mri_info
+from .._freesurfer import _read_mri_info, _check_mri
 from ..io import _loc_to_coil_trans
 from ..io.pick import pick_types, _picks_to_idx
 from ..io.constants import FIFF
@@ -41,7 +41,6 @@ from ..utils import (get_subjects_dir, logger, _check_subject, verbose, warn,
                      _ensure_int, _validate_type, _check_option)
 from .utils import (mne_analyze_colormap, _get_color_list,
                     plt_show, tight_layout, figure_nobar, _check_time_unit)
-from .misc import _check_mri
 from ..bem import (ConductorModel, _bem_find_surface,
                    read_bem_surfaces, _ensure_bem_surfaces)
